+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 <!-- markdownlint-disable no-trailing-punctuation -->
 
+## next
+
+TODO: Date
+
+- **Breaking changes**:
+  - Renamed trait `TipToed` to `RefCounted` and its function `tip_toe` to `ref_counter`.
+
 ## 0.0.1
 
 2021-11-08